set up a load balancer for vault.kitchen.horse
I'm not sure I like this model.
I like the forwarding rule -> http/https proxy -> backends (by rules)
thing, but I think the mapping of proxy -> rule -> backend should be an
external resource so I can make one load balancer for everything in
gcp-central and just attach things to it. This would save me from having
to have a bunch of external IPs that I have to pay for, or from having
to centralize all of my load balancer configuration